requirements.txt
PyQt6>=6.6.0
pyspellchecker>=0.7.2
